Move <ask> widgets to its own namespace
